feat: refactor transactions retrieval to separate method
